error no tests

+++ UserGiDataBase/src/main/java/com/minderaschool/UserGiDataBase/service/UserService.java
@@ -2,7 +2,9 @@ package com.minderaschool.UserGiDataBase.service;
 
 import com.minderaschool.UserGiDataBase.dto.UserDto;
 import com.minderaschool.UserGiDataBase.entity.UserEntity;
+import com.minderaschool.UserGiDataBase.exception.UserNotFoundException;
 import com.minderaschool.UserGiDataBase.repositoy.UserRepository;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -43,19 +45,21 @@ public class UserService {
         repository.save(existingUser);
     }
 
-    public void updatePassword(Integer id, UserDto updatePassword){
+    public void updatePatch(Integer id, UserDto updatePatch) {
         UserEntity user = repository.getReferenceById(id);
-        user.setPassword(updatePassword.getPassword());
+        if (updatePatch.getUsername() != null) {
+            user.setUsername(updatePatch.getUsername());
+        }
+        if (updatePatch.getPassword() != null) {
+            user.setPassword(updatePatch.getPassword());
+        }
         repository.save(user);
     }
 
-    public void updateUsername(Integer id, UserDto updateUsername){
-        UserEntity user = repository.getReferenceById(id);
-        user.setUsername(updateUsername.getUsername());
-        repository.save(user);
-    }
-
-    public void deleteUser(Integer id) {
+    public void deleteUser(Integer id) throws Exception {
+        if (repository.findById(id).isEmpty()) {
+            throw new UserNotFoundException("User "+id+" not found");
+        }
         repository.deleteById(id);
     }
 }
